Re-shape whole project
