+++ src/main.css
@@ -1,9 +1,32 @@
-#form-box {
+.form-box {
   width: 500px;
   margin: auto;
 }
 
-#spacingStyle {
-  display: block;
-  margin: 20px auto;
+.with-space {
+  display: block !important;
+  margin: 20px auto !important;
+}
+
+.MuiAlert-icon {
+  margin-top: 5.5px;
+}
+
+.input-hint {
+  font-family: "Roboto", sans-serif;
+  font-size: 12px;
+  margin-top: 3px;
+  margin-left: 2px;
+}
+
+.btn-text {
+  font-family: "Roboto", sans-serif;
+  font-size: 12px;
+  font-weight: 500;
+  border: none;
+  background-color: transparent;
+  margin: 0;
+  padding: 0;
+  text-decoration: underline;
+  cursor: pointer;
 }
