+++ src/main.tsx
@@ -1,15 +1,8 @@
-import { StrictMode } from 'react';
+import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
 import { ThemeProvider, createTheme } from '@mui/material';
-import {
-  EthereumClient,
-  w3mProvider,
-  w3mConnectors,
-} from '@web3modal/ethereum';
-import { WagmiConfig, createClient, configureChains } from 'wagmi';
-import { Web3Modal } from '@web3modal/react';
-import { bellecour } from './utils/walletConnection';
+import { WagmiConfig } from 'wagmi';
+import { wagmiConfig } from './utils/wagmiConfig.ts';
 import './main.css';
 import Front from './features/Front';
 
@@ -26,34 +19,12 @@ const theme = createTheme({
   },
 });
 
-// Wagmi Client
-if (!import.meta.env.VITE_WALLET_CONNECT_PROJECT_ID) {
-  throw new Error(
-    'You need to provide a WALLET_CONNECT_PROJECT_ID env variable'
-  );
-}
-const projectId = import.meta.env.VITE_WALLET_CONNECT_PROJECT_ID!;
-const chains = [bellecour];
-const { provider } = configureChains(chains, [w3mProvider({ projectId })]);
-
-const wagmiClient = createClient({
-  autoConnect: true,
-  connectors: w3mConnectors({ version: 1, chains, projectId }),
-  provider,
-});
-
-// Configure modal ethereum client
-const ethereumClient = new EthereumClient(wagmiClient, chains);
-
 root.render(
-  <StrictMode>
+  <React.StrictMode>
     <ThemeProvider theme={theme}>
-      <BrowserRouter>
-        <WagmiConfig client={wagmiClient}>
-          <Front />
-        </WagmiConfig>
-        <Web3Modal projectId={projectId} ethereumClient={ethereumClient} />
-      </BrowserRouter>
+      <WagmiConfig config={wagmiConfig}>
+        <Front />
+      </WagmiConfig>
     </ThemeProvider>
-  </StrictMode>
+  </React.StrictMode>
 );
