+++ src/utils/constants.ts
@@ -1,0 +1,2 @@
+//web3mail dapp END
+export const WEB3MAIL_APP_ENS = 'web3mail.apps.iexec.eth';
