+++ src/utils/types.ts
@@ -1,0 +1,3 @@
+export type Address = `0x${string}`;
+
+export type AddressOrEnsName = Address | string;
